исправлены ошибки2

+++ src/main.ts
@@ -5,7 +5,8 @@ import { CatalogProducts } from "./components/Models/CatalogProducts";
 import { ShopingCart } from "./components/Models/ShopingCart";
 import { apiProducts } from "./utils/data";
 import { Communication } from "./components/Models/Сommunication";
-import { IProduct } from "./types";
+import { Api } from "./components/base/Api";
+import { API_URL } from "./utils/constants";
 
 const products = new CatalogProducts();
 products.setCatalog(apiProducts.items);
@@ -14,18 +15,23 @@ console.log(
   "Найденный по id товар: ",
   products.getProductFromId("b06cde61-912f-4663-9751-09956c0eed67"),
 );
-products.setCurrentProduct(
-  products.getProductFromId("b06cde61-912f-4663-9751-09956c0eed67") as IProduct,
+const product1 = products.getProductFromId(
+  "b06cde61-912f-4663-9751-09956c0eed67",
 );
+if (product1) products.setCurrentProduct(product1);
 console.log(`Выбранный товар: `, products.getCurrentProduct());
 
 const cart = new ShopingCart();
-cart.addProduct(
-  products.getProductFromId("412bcf81-7e75-4e70-bdb9-d3c73c9803b7") as IProduct,
+const product2 = products.getProductFromId(
+  "412bcf81-7e75-4e70-bdb9-d3c73c9803b7",
 );
-cart.addProduct(
-  products.getProductFromId("854cef69-976d-4c2a-a18c-2aa45046c390") as IProduct,
+const product3 = products.getProductFromId(
+  "854cef69-976d-4c2a-a18c-2aa45046c390",
 );
+if (product2 && product3) {
+  cart.addProduct(product2);
+  cart.addProduct(product3);
+}
 console.log("Что в корзине: ", cart.getContents());
 console.log("Стоимость товаров в корзине: ", cart.getPriceProducts());
 console.log("Количество товаров в корзине: ", cart.getQuantity());
@@ -33,9 +39,7 @@ console.log(
   "В корзине ли товар? : ",
   cart.isInTheCart("412bcf81-7e75-4e70-bdb9-d3c73c9803b7"),
 );
-cart.deleteProduct(
-  products.getProductFromId("854cef69-976d-4c2a-a18c-2aa45046c390") as IProduct,
-);
+if (product3) cart.deleteProduct(product3);
 console.log("Что в корзине после удаления 1 элемента: ", cart.getContents());
 cart.cleanCart();
 console.log("Что в корзине после очистки: ", cart.getContents());
@@ -55,16 +59,15 @@ console.log(`Данные заполненной формы: `, customer.getData
 customer.cleanData();
 console.log(`Ошибки формы после очистки данных: `, customer.validateData());
 
-const api = new Communication();
+const apiBase = new Api(API_URL);
+const api = new Communication(apiBase);
 const testProducts = new CatalogProducts();
-async function init() {
-  const res = await api.get();
-  return res;
-}
-
-init().then(res => {
-    testProducts.setCatalog(res);
-    console.log(`Выводим полученные с сервера товары: `, testProducts.getCatalog());
-  }).catch(console.error);
-
-
+api.getProducts()
+  .then((res) => {
+    testProducts.setCatalog(res.items);
+    console.log(
+      `Выводим полученные с сервера товары: `,
+      testProducts.getCatalog(),
+    );
+  })
+  .catch(console.error);
